[passo2] extracao do routes
